+++ src/index.ts
@@ -1,182 +1,192 @@
 import {
-    ICommandPalette
-} from '@jupyterlab/apputils';
+    ICommandPalette,
+} from "@jupyterlab/apputils";
 
 import {
-    JupyterLab, JupyterLabPlugin
-} from '@jupyterlab/application';
+    JupyterLab, JupyterLabPlugin,
+} from "@jupyterlab/application";
 
 import {
-    INotebookTracker
-} from '@jupyterlab/notebook';
+    INotebookTracker,
+} from "@jupyterlab/notebook";
 
 import {
-    Terminal
-} from '@jupyterlab/terminal';
+    Terminal,
+} from "@jupyterlab/terminal";
 
 import {
-    CodeCell
-} from '@jupyterlab/cells';
+    CodeCell,
+} from "@jupyterlab/cells";
 
 import {
     ISettingRegistry,
-} from '@jupyterlab/coreutils';
+} from "@jupyterlab/coreutils";
 
-import '../style/index.css';
+import "../style/index.css";
 
 // NOTE: EOL for when we write to tmp, there probably
 //       isn't a chance of user's input to colide with this...
-const EOL = 'I97Qo3mjGJLySbuljz1Fzvh1ulPWBQdkG0QZ';
+const EOL = "I97Qo3mjGJLySbuljz1Fzvh1ulPWBQdkG0QZ";
 // NOTE: Listening to a terminal and determining which message is meant for
 //       us is hard, let's make this less hard by prepending some string
-const STDOUT_MARKER = 'Cl3PsCL5Ionv1d9tQMvNXRtz5aXUHj2VMk3mnn7Sg179';
-
+const STDOUT_MARKER = "Cl3PsCL5Ionv1d9tQMvNXRtz5aXUHj2VMk3mnn7Sg179";
 
 class JupyterLabBlackFormatter {
-    _app: JupyterLab;
-    _tracker: INotebookTracker;
-    _palette: ICommandPalette;
-    _term: Terminal;
-    _settingRegistry: ISettingRegistry;
+    private app: JupyterLab;
+    private tracker: INotebookTracker;
+    private palette: ICommandPalette;
+    private term: Terminal;
+    private settingRegistry: ISettingRegistry;
 
-    _loaded = false;
-    _working = false;
-    _success = false;
-    _lineLength: number;
-    _python: string;
+    private loaded = false;
+    private working = false;
+    private success = false;
+    private lineLength: number;
+    private python: string;
 
-    constructor(app: JupyterLab, tracker: INotebookTracker, palette: ICommandPalette, settingRegistry: ISettingRegistry){
-        this._app = app;
-        this._tracker = tracker;
-        this._palette = palette;
-        this._settingRegistry = settingRegistry;
+    constructor(
+        app: JupyterLab, tracker: INotebookTracker,
+        palette: ICommandPalette, settingRegistry: ISettingRegistry,
+    ) {
+        this.app = app;
+        this.tracker = tracker;
+        this.palette = palette;
+        this.settingRegistry = settingRegistry;
         this.setupButton();
         this.setupTerminalSession();
         this.setupSettings();
     }
 
-    setupSettings(){
-        let self = this;
-        Promise.all([this._settingRegistry.load('@ryantam626/jupyterlab_black:settings')]).then(
+    private setupSettings() {
+        const self = this;
+        Promise.all([this.settingRegistry.load("@ryantam626/jupyterlab_black:settings")]).then(
             ([settings]) => {
-                function onSettingsUpdated(settings: ISettingRegistry.ISettings){
-                    self._lineLength = settings.get('lineLength').composite as number;
-                    self._python = settings.get('blackPythonBin').composite as string;
+                function onSettingsUpdated(jsettings: ISettingRegistry.ISettings) {
+                    self.lineLength = jsettings.get("lineLength").composite as number;
+                    self.python = jsettings.get("blackPythonBin").composite as string;
                 }
                 settings.changed.connect(onSettingsUpdated);
                 onSettingsUpdated(settings);
-            }
+            },
+        // tslint:disable-next-line:no-console
         ).catch((reason: Error) => console.error(reason.message));
     }
 
     // TODO: Attempt to reuse a terminal somewhere..
 
-    async setupTerminalSession(){
+    private async setupTerminalSession() {
         // Abort if somehow no terminals available.
-        if (!this._app.serviceManager.terminals.isAvailable()) {
-            console.log('Disabling jupyterlab_black plugin because lack of terminal access.');
-            this._loaded = false;
+        if (!this.app.serviceManager.terminals.isAvailable()) {
+            // tslint:disable-next-line:no-console
+            console.log("Disabling jupyterlab_black plugin because lack of terminal access.");
+            this.loaded = false;
             return;
         }
 
-        this._term = new Terminal();
+        this.term = new Terminal();
 
         try {
-            this._term.session = await this._app.serviceManager.terminals.startNew();
+            this.term.session = await this.app.serviceManager.terminals.startNew();
             // NOTE: unset HISTFILE so user's shell history is no longer as polluted
             //       when we dump codecell content though terminal.
-            this._term.session.send({type: 'stdin', content: ['unset HISTFILE\r']})
-            this._loaded = true;
-            console.log('Terminal session started.', this._term.session.name);
-        } catch(e) {
-            this._term.dispose();
-            this._term = undefined;
-            this._loaded = false;
+            this.term.session.send({type: "stdin", content: ["unset HISTFILE\r"]});
+            this.loaded = true;
+            // tslint:disable-next-line:no-console
+            console.log("Terminal session started.", this.term.session.name);
+        } catch (e) {
+            this.term.dispose();
+            this.term = undefined;
+            this.loaded = false;
         }
     }
 
-    maybeFormatCodecell(){
+    private maybeFormatCodecell() {
         // TODO: Check current kernel is of Python2/3
-        if (!this._loaded) {
-            console.log('Terminal session is NOT started. Retry??');
-        } else if (this._working) {
-            console.log('Already working on something!! CHILL.');
+        if (!this.loaded) {
+            // tslint:disable-next-line:no-console
+            console.log("Terminal session is NOT started. Retry??");
+        } else if (this.working) {
+            // tslint:disable-next-line:no-console
+            console.log("Already working on something!! CHILL.");
         } else {
-            if (this._tracker.activeCell instanceof CodeCell) {
-                this._working = true;
-                this._success = false;
-                this.dump_to_tmp(this._tracker.activeCell.model.value.text);
-                if (this.format_tmp(this._tracker.activeCell.model.id)) {
-                    this.retrieve_and_set_formatted_code(this._tracker.activeCell);
+            if (this.tracker.activeCell instanceof CodeCell) {
+                this.working = true;
+                this.success = false;
+                this.dump_to_tmp(this.tracker.activeCell.model.value.text);
+                if (this.format_tmp(this.tracker.activeCell.model.id)) {
+                    this.retrieve_and_set_formatted_code(this.tracker.activeCell);
                 }
             } else {
+                // tslint:disable-next-line:no-console
                 console.log("This doesn't seem like a code cell...");
             }
         }
     }
 
-    dump_to_tmp(content: string){
-        this._term.session.send({
-            type: 'stdin',
+    private dump_to_tmp(content: string) {
+        this.term.session.send({
             content: [
 `cat > /tmp/jupyterlab_black_working.py <<${EOL}
 ${content}
-${EOL}\r`
-            ]
+${EOL}\r`,
+            ],
+            type: "stdin",
         });
     }
 
-    format_tmp(codeCellIdent: string){
-        if (this._python === undefined || this._python === null) {
-            console.error('It appears the python interpreter is not set!');
-            this._working = false;
+    private format_tmp(codeCellIdent: string) {
+        if (this.python === undefined || this.python === null) {
+            // tslint:disable-next-line:no-console
+            console.error("It appears the python interpreter is not set!");
+            this.working = false;
             return false;
         }
-        this._term.session.send({
-            type: 'stdin',
+        this.term.session.send({
             content: [
-`${this._python} - <<${EOL}
+`${this.python} - <<${EOL}
 from black import format_str
 with open('/tmp/jupyterlab_black_working.py', 'r') as file_:
     unformatted = file_.read()
-formatted = format_str(unformatted, line_length=${this._lineLength})
+formatted = format_str(unformatted, line_length=${this.lineLength})
 with open('/tmp/jupyterlab_black_working.py', 'w') as file_:
     file_.write("${STDOUT_MARKER}${codeCellIdent}" + formatted)
-${EOL}\r`
-            ]
+${EOL}\r`,
+            ],
+            type: "stdin",
+
         });
         return true;
     }
 
-    retrieve_and_set_formatted_code(activeCell: CodeCell){
-        let self = this;
+    private retrieve_and_set_formatted_code(activeCell: CodeCell) {
+        const self = this;
 
-        function cleanTerminalOutput(termOutput: string): string{
-            let cleanedOutput = termOutput.replace(STDOUT_MARKER + activeCell.model.id, '');
+        function cleanTerminalOutput(termOutput: string): string {
+            let cleanedOutput = termOutput.replace(STDOUT_MARKER + activeCell.model.id, "");
             // Convert line endings if necessary
             // TOOD: Find out what happens with Mac/Windows...
-            cleanedOutput = cleanedOutput.replace(/\r\n|\r/g, '\n');
+            cleanedOutput = cleanedOutput.replace(/\r\n|\r/g, "\n");
             // NOTE: We don't want a newline at the end of the code cell!
             cleanedOutput = cleanedOutput.substr(0, cleanedOutput.length - 1);
             return cleanedOutput;
         }
 
-        function onTerminalMessage(sender: any, rawMessage: any): void{
-            let message: string = rawMessage.content[0];
-            if (message.indexOf(STDOUT_MARKER + activeCell.model.id) > -1){
+        function onTerminalMessage(sender: any, rawMessage: any): void {
+            const message: string = rawMessage.content[0];
+            if (message.indexOf(STDOUT_MARKER + activeCell.model.id) > -1) {
                 activeCell.model.value.text = cleanTerminalOutput(message);
-                self._success = true;
-                self._term.session.messageReceived.disconnect(onTerminalMessage, self);
+                self.success = true;
+                self.term.session.messageReceived.disconnect(onTerminalMessage, self);
             }
         }
-        this._term.session.messageReceived.connect(onTerminalMessage, this);
+        this.term.session.messageReceived.connect(onTerminalMessage, this);
         // FIXME: This probably doesn't work for Windows?
-        this._term.session.send({type: 'stdin', content: ['cat /tmp/jupyterlab_black_working.py\r']})
+        this.term.session.send({type: "stdin", content: ["cat /tmp/jupyterlab_black_working.py\r"]});
 
-        let check_success_promise = new Promise((resolve, reject) => {
-            (function check_success(){
-                setTimeout(function(){
-                    if (self._success){
+        const checkSuccessPromise = new Promise((resolve, reject) => {
+            (function check_success() {
+                setTimeout(() =>  {
+                    if (self.success) {
                         resolve();
                     } else {
                         check_success();
@@ -184,28 +194,29 @@ ${EOL}\r`
                 }, 50);
             })();
         });
-        let check_failure_promise = new Promise(function(resolve, reject) {
+        const checkFailurePromise = new Promise((resolve, reject) => {
             setTimeout(reject, 1000);
         });
 
-        Promise.race([check_success_promise, check_failure_promise]).then(response => {
-            this._working = false;
-        }).catch(response => {
-            this._term.session.messageReceived.connect(onTerminalMessage, this);
-            this._working = false;
-            console.error('Something went wrong, check your python interpreter settings is correct maybe?')
+        Promise.race([checkSuccessPromise, checkFailurePromise]).then((response) => {
+            this.working = false;
+        }).catch((response) => {
+            this.term.session.messageReceived.connect(onTerminalMessage, this);
+            this.working = false;
+            // tslint:disable-next-line:no-console
+            console.error("Something went wrong, check your python interpreter settings is correct maybe?");
         });
     }
 
-    setupButton(){
+    private setupButton() {
         const command = "jupyterlab_black:format";
-        this._app.commands.addCommand(command,{
-            label: "Apply Black Formatter",
+        this.app.commands.addCommand(command, {
             execute: () => {
                 this.maybeFormatCodecell();
-            }
+            },
+            label: "Apply Black Formatter",
         });
-        this._palette.addItem( {command, category: "JupyterLab Black"} );
+        this.palette.addItem( {command, category: "JupyterLab Black"} );
     }
 
 }
@@ -214,12 +225,16 @@ ${EOL}\r`
  * Initialization data for the jupyterlab_black extension.
  */
 const extension: JupyterLabPlugin<void> = {
-  id: 'jupyterlab_black',
-  autoStart: true,
-  requires: [ICommandPalette, INotebookTracker, ISettingRegistry],
-  activate: (app: JupyterLab, palette: ICommandPalette, tracker: INotebookTracker, settingRegistry: ISettingRegistry) => {
-    new JupyterLabBlackFormatter(app, tracker, palette, settingRegistry);
-  }
+    activate: (
+        app: JupyterLab, palette: ICommandPalette,
+        tracker: INotebookTracker, settingRegistry: ISettingRegistry,
+    ) => {
+        // tslint:disable-next-line:no-unused-expression
+        new JupyterLabBlackFormatter(app, tracker, palette, settingRegistry);
+    },
+    autoStart: true,
+    id: "jupyterlab_black",
+    requires: [ICommandPalette, INotebookTracker, ISettingRegistry],
 };
 
 export default extension;