+++ CHANGELOG.md
@@ -4,7 +4,7 @@ UNRELEASED
 * Bump dependencies ranges;
 * Add note about keyboard shortcut;
 * Add badges to README;
-
+* Apply recommended tslint rules;
 
 0.1.3 2018-07-19
 ----------------